refactor: move collectCargo() method to separate module
